refactor(qtensor): dispatch functions per subclass
