+++ optimum/quanto/tensor/qbits/qbits.py
@@ -247,6 +247,31 @@ class QBitsTensor(QTensor):
         return QBitsTensor(qtype, axis, group_size, size, stride, data, scale, shift)
 
     @classmethod
+    def __torch_function__(cls, func, types, args=(), kwargs=None):
+        """Dispatch torch functions applied on this subtensor
+
+        This method is called whenever a torch function (such as `torch.nn.functional.linear`)
+        is called with at least one parameter coresponding to this subtensor:
+
+        - if a quantized implementation exists for the selected function, it is called,
+        - otherwise, the original implementation is called, deactivating further functional dispatch.
+
+        During the execution of the standard torch function, a second-level of dispatch will
+        happen, but this time directly on individual torch Tensor operations (mainly ATEN).
+        """
+        from ..qtensor_func import get_qtensor_func
+
+        kwargs = kwargs or {}
+
+        # Look for a func accepting QTensor inputs
+        qfunc = get_qtensor_func(func)
+        if qfunc is not None:
+            return qfunc(*args, **kwargs)
+        # Defer to dispatcher to look instead for QTensor subclasses operations
+        with torch._C.DisableTorchFunctionSubclass():
+            return func(*args, **kwargs)
+
+    @classmethod
     def __torch_dispatch__(cls, op, types, args, kwargs=None):
         from .qbits_ops import get_qbitstensor_op_dispatch
