+++ optimum/quanto/tensor/qbits/tinygemm/qbits.py
@@ -125,3 +125,28 @@ class TinyGemmQBitsTensor(QBitsTensor):
         size = ast.literal_eval(meta["size"])
         stride = ast.literal_eval(meta["stride"])
         return TinyGemmQBitsTensor(qtype, axis, group_size, size, stride, data, scale_shift)
+
+    @classmethod
+    def __torch_function__(cls, func, types, args=(), kwargs=None):
+        """Dispatch torch functions applied on this subtensor
+
+        This method is called whenever a torch function (such as `torch.nn.functional.linear`)
+        is called with at least one parameter coresponding to this subtensor:
+
+        - if a quantized implementation exists for the selected function, it is called,
+        - otherwise, the original implementation is called, deactivating further functional dispatch.
+
+        During the execution of the standard torch function, a second-level of dispatch will
+        happen, but this time directly on individual torch Tensor operations (mainly ATEN).
+        """
+        from ...qtensor_func import get_qtensor_func
+
+        kwargs = kwargs or {}
+
+        # Look for a func accepting QTensor inputs
+        qfunc = get_qtensor_func(func)
+        if qfunc is not None:
+            return qfunc(*args, **kwargs)
+        # Defer to dispatcher to look instead for QTensor subclasses operations
+        with torch._C.DisableTorchFunctionSubclass():
+            return func(*args, **kwargs)
